+++ src/tools/bt-type.py
@@ -6,7 +6,6 @@
 
 
 import argparse
-import os
 import sys
 
 from libs import btklib
@@ -15,9 +14,10 @@ from libs import constants
 
 def main(args):
     message = args.message
+    delay = args.delay
 
     btk = btklib.BtkLib()
-    btk.send_text(message, delay=constants.TENTH_OF_A_SECOND)
+    btk.send_text(message, delay=delay)
 
 
 # argument parser
@@ -25,8 +25,11 @@ def __parse_arguments(args):
 
     parser = argparse.ArgumentParser(description="Type using BT keyboard")
 
-    parser.add_argument("-m", "--message", required=True,
+    parser.add_argument("message", 
                         help="Message to type")
+    
+    parser.add_argument("-d", "--delay", type=float, default=constants.BT_IOS_COMMANDS_DEFAULT_KEYBOARD_TYPE_DELAY,
+                        help=f"Delay between characters (in seconds). Default is {constants.BT_IOS_COMMANDS_DEFAULT_KEYBOARD_TYPE_DELAY}.")
 
     parsed = parser.parse_args(args)
     return parsed