+++ src/tools/libs/automation/lighthouse_commands.py
@@ -11,21 +11,21 @@ def activate_port_fw(device, port=9222):
 
     command = f"adb -s {device['adb_identifier']} forward tcp:{port} localabstract:chrome_devtools_remote"
     os.system(command)
-    time.sleep(constants.THREE_SECONDS)
+    time.sleep(constants.LIGHTHOUSE_COMMANDS_WAIT_TIME)
 
 
 def deactivate_port_fw(device, port=9222):
 
     command = f"adb -s {device['adb_identifier']} forward --remove tcp:{port}"
     os.system(command)
-    time.sleep(constants.THREE_SECONDS)
+    time.sleep(constants.LIGHTHOUSE_COMMANDS_WAIT_TIME)
 
 
 def deactivate_all_port_fw(device):
 
     command = f"adb -s {device['adb_identifier']} forward --remove-all"
     os.system(command)
-    time.sleep(constants.THREE_SECONDS)
+    time.sleep(constants.LIGHTHOUSE_COMMANDS_WAIT_TIME)
 
 
 def measure_url(url, path, port=9222, max_wait_load_time=200000):
@@ -39,4 +39,4 @@ def measure_url(url, path, port=9222, max_wait_load_time=200000):
 
     command = f"lighthouse {url} --max-wait-for-load {max_wait_load_time} --port={port} --save-assets --emulated-form-factor=none --throttling-method=provided --output-path={filepath} --output=json"
     os.system(command)
-    time.sleep(constants.THREE_SECONDS)
+    time.sleep(constants.LIGHTHOUSE_COMMANDS_WAIT_TIME)